remove python demo and consolidate
