--- demos/realtime_motion_graph/__init__.py
@@ -1,1 +1,0 @@
-"""Realtime motion-to-music demo (full demo, server, and thin client)."""
