--- demos/realtime_motion_graph/client/__main__.py
@@ -1,4 +1,0 @@
-from .app import main
-
-if __name__ == "__main__":
-    main()
