+++ demos/realtime_motion_graph_web/__init__.py
@@ -1,5 +1,5 @@
 """Web-app version of the realtime motion-to-music demo.
 
-Serves a browser-based client over HTTP while reusing the exact same
-GPU pipeline as :mod:`demos.realtime_motion_graph.server`.
+Serves a browser-based client over HTTP and runs the GPU pipeline on
+the same TCP port. Entry point: ``python -m demos.realtime_motion_graph_web``.
 """
